moved result influxdb methods to a trait
